GPS and SD card data logger
Ubloc Neo-6m GPS module is used  - Pin 2 and 3 is as TX and RX reader
Mirco SD card adapter module is connected with the following Pin IDs Vcc to 5v, MISO to pin 12 (50), MOSI to Pin 11 (51), SCK to pin 13 (52) and CS to Pin 4 (53). Note for Arduino UNO use Pin ID without brackets and for Mega use Pin ID with brackets.